requirements.txt
astor==0.8.1
black==22.6.0
click==8.1.3
colorama==0.4.5
flynt==0.76
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyYAML==6.0
tomli==2.0.1
typing-extensions==4.3.0
